// PRO-EP/pro-ep-ui/src/app/interfaces/interfaces.ts
export interface IUserMin {
  userid: number;
  dempoid: string;
  displayName: string | null;
  fname: string;
  lname: string;
  preferredfname: string;
  preferredlname: string;
  language: string;
  role: number;
  buddy?: boolean;
  employmenttype?: number;
  trainedon: string;
  trainedOnArray: string[];
  active?: boolean;
  canEdit?: boolean;
  selected?: boolean;
  checked?: boolean;
  changed?: boolean;
  schedulinglevel?: number;
}

export interface ITimeCard {
  timeCardId: number;
  dempoid: string;
  //workDate?: Date;
  timeIn?: Date | null;
  timeOut?: Date | null;
  entryBy: string;
  entryDt?: Date;
  modBy: string;
  modDt?: Date;
  datetimein?: Date | null;
  datetimeout?: Date | null;
  externalIp: string | null;
  //MachineName: string;
}
// ...
export interface IProjectMin {
  projectID: number;
  projectName: string;
  projectAbbr: string;
  projectDisplayId: string;
  displayedIn?: string;
  active?: boolean | null;
  projectStatus?: boolean | null;
  projectColor: string;
  projectType?: string | null;
  selected?: boolean;
  trainedOnUsersMin?: IUserMin[];
  trainedOnUsers?: string[];
  notTrainedOnUsers?: string[];
}
// ...
export interface ISchedule {
  preschedulekey: number;
  dempoid: string | null;
  fname: string | null;
  lname: string | null;
  preferredfname: string | null;
  preferredlname: string | null;
  displayName: string | null;
  userName: string | null;
  projectid?: number | null;
  projectName: string | null;
  projectColor: string;
  scheduledate: Date;
  comments: string | null;
  startdatetime?: Date | null;
  enddatetime?: Date | null;
  startTime?: string;
  endTime?: string;
  dayNumber?: number | null;
  weekNumber?: number | null;
  weekStart: Date | null;
  weekEnd: Date | null;
  dayOfWeek: number;
  dayOfWeekText?: string;
  monthNumber?: number | null;
  month: string | null;
  scheduleyear?: number | null;
  scheduledHours?: number | null;
  expr1: string | null;
  requestId?: number | null;
  requestDetails: string | null;
  requestCodeId?: number | null;
  requestCode: string | null;
  userid: number | null;
  trainedon: string | null;
  language: string | null;
  changed?: boolean;
  saved?: boolean | null;
  markedForDeletion?: boolean | null;
  addTab?: boolean;
  addInitial?: boolean;
  invalid?: boolean;
  invalidFields?: string[];
  entryBy: string | null;
  entryDt?: Date;
  validationMessages?: string[];
  hoverMessage?: string;
  scheduleConflict?: boolean;
  justAdded?: boolean;
  initialProjectid?: number | null;
}

export interface IScheduleMin {
  preschedulekey: number;
  dempoid: string | null;
  scheduledate?: Date;
  startdatetime?: Date | null;
  enddatetime?: Date | null;
  status?: number;
  projectid?: number | null;
  comments: string | null;
  entryBy: string;
  entryDt?: Date;
  modBy: string;
  modDt?: Date;
  machineName: string;
}
// ...
export interface ITimeCode {
  timePeriod: string;
  timeCodeValue: number;
  timePeriodampm: string;
}

export interface IAnyFilter {
  filterControl: FormControl;
  anyToggle: boolean;
}

export interface ISaveResponse {
  status: string;
  successCount: number;
  errors: Array<string>;
}

export interface IGeneralResponse {
  Status: string;
  Message: string;
  Subject: any;
}

export interface IAuthenticatedUser {
  displayName: string;
  duDukeID: string;
  eppn: string;
  netID: string;
  isMemberOf: string;
  interviewer: boolean;
  resourceGroup: boolean;
  admin: boolean;
  userRoles?: string[] | undefined;
  sessionMinsLeft: number;
  timecards: ITimeCard[];
}

export interface IDateRange {
  startDate: Date;
  endDate: Date;
}


export interface IUserTrainedOn {
  netId: string;
  trainedOn: string[];
}

export interface IPopupMessage {
  title: string;
  body: string;
  buttons: IActionButton[] | null;
  closeCallback: Function | null;
  noBlackFilter?: boolean;
}

export interface IActionButton {
  label: string;
  callbackFunction?: Function;
}
// ...
export interface IProjectGroup {
  name: string;
  projects: IProjectMin[];
}

export interface IValidationMessage {
  validationMessagesId: number;
  dempoId: string | null;
  messageId: number;
  inMonth: Date;
  scheduleKeys: string | null;
  messageText: string | null;
  details: string | null;
  schedules?: ISchedule[];
}

export interface IValidationMessageText {
  validationMessageTextId: number;
  messageId: number;
  orderId: number;
  messageText: string;
}

export interface IBlockOutDate {
  blockOutDateId: number;
  blockOutDay: Date | null;
  allDay?: boolean;
  startTime: string;
  endTime: string;
}